// frontend-nextjs/src/api/generated/api.ts
export interface ExpenseDto {
    /**
     * 家計簿データの一意ID
     * @type {number}
     * @memberof ExpenseDto
     */
    'id'?: number;
    /**
     * 支出日
     * @type {string}
     * @memberof ExpenseDto
     */
    'date'?: string;
    /**
     * 支出カテゴリ
     * @type {string}
     * @memberof ExpenseDto
     */
    'category'?: string;
    /**
     * 支出金額
     * @type {number}
     * @memberof ExpenseDto
     */
    'amount'?: number;
    /**
     * 支出の詳細説明
     * @type {string}
     * @memberof ExpenseDto
     */
    'description'?: string;
}